fix: fix csrf token in local
