+++ shell/app/locales/en.json
@@ -288,7 +288,6 @@
     "Status-alert": "Status",
     "Submit time": "Submit time",
     "Subscription": "Subscription",
-    "Supports agile management of Scrum and traditional project management, including requirements, tasks and bugs.": "Supports agile management of Scrum and traditional project management, including requirements, tasks and bugs.",
     "System load": "System load",
     "Template Sample": "Template Sample",
     "The allocated node should be a certain one in the node list. Please check.": "The allocated node should be a certain one in the node list. Please check.",
@@ -307,7 +306,6 @@
     "Unable to go offline normally": "Unable to go offline normally",
     "Upload File": "Upload File",
     "Usage": "Usage",
-    "Used to host code repositories, supports multiple repositories, with the option to enable CI/CD and artifacts.": "Used to host code repositories, supports multiple repositories, with the option to enable CI/CD and artifacts.",
     "VIP segment": "VIP segment",
     "VPC": "VPC",
     "VPC label": "VPC label",
