+++ shell/app/locales/zh.json
@@ -288,7 +288,6 @@
     "Status-alert": "告警状态",
     "Submit time": "发起申请时间",
     "Subscription": "包年包月",
-    "Supports agile management of Scrum and traditional project management, including requirements, tasks and bugs.": "支持Scrum敏捷管理和传统任务管理项目，包含需求、任务和缺陷。",
     "System load": "系统负载",
     "Template Sample": "模板样例",
     "The allocated node should be a certain one in the node list. Please check.": "分配节点应为节点列表中的某一个，请核对",
@@ -307,7 +306,6 @@
     "Unable to go offline normally": "无法正常下线原因",
     "Upload File": "上传文件",
     "Usage": "使用率",
-    "Used to host code repositories, supports multiple repositories, with the option to enable CI/CD and artifacts.": "用于托管代码仓库，支持多个仓库，并可选择开启CI/CD和制品。",
     "VIP segment": "VIP网段",
     "VPC": "专有网络",
     "VPC label": "VPC 标签",
